Update storage-format-(mac).txt
Fixed the instructions for the mac part as they were not correct for vita use. We need to use the "sudo dd if=/dev/zero of=/dev/disk2" to correctly clean the microsd (just like for the linux part) which mac os doesn't do by default with it's disk utility.