Added more instruction for non memorycard users
It is better to make them create an internal memory right away. When they will downgrade, vitadeploy will disappear if installed in near, and running Henlo again will not work as they would need to disable spoof so that Henlo correctly detect the version they are in and use the correct ropchain for the hack.