Kang publish script from dsi.cfw.guide
The VuePress 2.0.0-b49 changes were only tested with npm. It is unknown
whether yarn will actually work with the changes.

dsi.cfw.guide already uses npm for this, so let's just use the script
from there as it is known to work.

This file was taken from the following snapshot:
https://github.com/cfw-guide/dsi.cfw.guide/blob/66d1089257a7e6bda97ed337d9658fc3c836f656/.github/workflows/publish.yml